+++ core/render/vi/__init__.py
@@ -1,13 +1,11 @@
-from viur.core.prototypes import List
+# noinspection PyUnresolvedReferences
+from viur.core.render.vi.user import UserRender as user  # this import must exist!
 from viur.core.render.json.default import DefaultRender, CustomJsonEncoder
-from viur.core.render.vi.user import UserRender as user
-from viur.core import conf, exposed
-from viur.core import securitykey
-from viur.core import utils
-from viur.core import errors
-import datetime, json
+from viur.core import Module, conf, exposed, securitykey, utils, errors
 from viur.core.utils import currentRequest, currentLanguage
 from viur.core.skeleton import SkeletonInstance
+import datetime
+import json
 
 
 class default(DefaultRender):
@@ -17,111 +15,102 @@ class default(DefaultRender):
 __all__ = [default]
 
 
+@exposed
 def genSkey(*args, **kwargs):
+    currentRequest.get().response.headers["Content-Type"] = "application/json"
     return json.dumps(securitykey.create())
 
 
-genSkey.exposed = True
-
-
+@exposed
 def timestamp(*args, **kwargs):
     d = datetime.datetime.now()
+    currentRequest.get().response.headers["Content-Type"] = "application/json"
     return json.dumps(d.strftime("%Y-%m-%dT%H-%M-%S"))
 
 
-timestamp.exposed = True
-
-
-def getStructure(adminTree, module):
-    if not module in dir(adminTree) \
-        or not "adminInfo" in dir(getattr(adminTree, module)) \
-        or not getattr(adminTree, module).adminInfo:
-        # Module not known or no adminInfo for that module
+@exposed
+def getStructure(module):
+    """
+    Returns all available skeleton structures for a given module.
+    """
+    moduleObj = getattr(conf["viur.mainApp"].vi, module, None)
+    if not isinstance(moduleObj, Module) or not moduleObj.describe():
         return json.dumps(None)
+
     res = {}
-    try:
-        moduleObj = getattr(adminTree, module)
-    except:
-        return None
-    for stype in ["viewSkel", "editSkel", "addSkel", "viewLeafSkel", "viewNodeSkel", "editNodeSkel", "editLeafSkel",
-                  "addNodeSkel", "addLeafSkel"]:  # Unknown skel type
-        if stype in dir(moduleObj):
-            try:
-                skel = getattr(moduleObj, stype)()
-            except (TypeError, ValueError):
-                continue
-            if isinstance(skel, SkeletonInstance):
-                res[stype] = default().renderSkelStructure(skel)
-    if not res and "nodeSkelCls" in dir(moduleObj):
+
+    # check for tree prototype
+    if "nodeSkelCls" in dir(moduleObj):
         # Try Node/Leaf
-        for stype in ["viewSkel", "editSkel", "addSkel"]:
-            for treeType in ["node", "leaf"]:
+        for stype in ("viewSkel", "editSkel", "addSkel"):
+            for treeType in ("node", "leaf"):
                 if stype in dir(moduleObj):
                     try:
                         skel = getattr(moduleObj, stype)(treeType)
                     except (TypeError, ValueError):
                         continue
+
                     if isinstance(skel, SkeletonInstance):
                         storeType = stype.replace("Skel", "") + ("LeafSkel" if treeType == "leaf" else "NodeSkel")
                         res[storeType] = default().renderSkelStructure(skel)
+    else:
+        # every other prototype
+        for stype in ("viewSkel", "editSkel", "addSkel"):  # Unknown skel type
+            if stype in dir(moduleObj):
+                try:
+                    skel = getattr(moduleObj, stype)()
+                except (TypeError, ValueError):
+                    continue
+                if isinstance(skel, SkeletonInstance):
+                    res[stype] = default().renderSkelStructure(skel)
 
     currentRequest.get().response.headers["Content-Type"] = "application/json"
-    if res:
-        return json.dumps(res, cls=CustomJsonEncoder)
-    else:
-        return json.dumps(None)
+    return json.dumps(res or None, cls=CustomJsonEncoder)
 
 
+@exposed
 def setLanguage(lang, skey):
     if not securitykey.validate(skey):
-        return
+        raise errors.PreconditionFailed()
+
     if lang in conf["viur.availableLanguages"]:
         currentLanguage.set(lang)
 
 
-setLanguage.exposed = True
+@exposed
+def dumpConfig():
+    res = {}
 
+    for key in dir(conf["viur.mainApp"].vi):
+        module = getattr(conf["viur.mainApp"].vi, key, None)
+        if not isinstance(module, Module):
+            continue
 
-def dumpConfig(adminTree):
-    adminConfig = {}
-    for key in dir(adminTree):
-        app = getattr(adminTree, key)
-        if "adminInfo" in dir(app) and app.adminInfo:
-            if callable(app.adminInfo):
-                info = app.adminInfo()
-                if info is not None:
-                    adminConfig[key] = info
-            else:
-                adminConfig[key] = app.adminInfo.copy()
-                adminConfig[key]["name"] = str(adminConfig[key]["name"])
-                adminConfig[key]["views"] = []
-                if "views" in app.adminInfo:
-                    for v in app.adminInfo["views"]:
-                        tmp = v.copy()
-                        tmp["name"] = str(tmp["name"])
-                        adminConfig[key]["views"].append(tmp)
-        if key in adminConfig:
-            adminConfig[key]["indexes"] = getattr(app, "indexes", [])
+        if admin_info := module.describe():
+            res[key] = admin_info
 
     res = {
-        "modules": adminConfig,
-        "configuration": {}
+        "modules": res,
+        "configuration": {
+            k.removeprefix("admin."): v for k, v in conf.items() if k.lower().startswith("admin.")
+        }
     }
-
-    for k, v in conf.items():
-        if k.lower().startswith("admin."):
-            res["configuration"][k[6:]] = v
 
     currentRequest.get().response.headers["Content-Type"] = "application/json"
     return json.dumps(res)
 
 
+@exposed
 def getVersion(*args, **kwargs):
-    # We force the patch-level of our version to be always zero for security reasons
+    """
+    Returns viur-core version number
+    """
+    currentRequest.get().response.headers["Content-Type"] = "application/json"
+    if conf["viur.instance.is_dev_server"]:
+        return json.dumps(conf["viur.version"])
+
+    # Hide patchlevel
     return json.dumps((conf["viur.version"][0], conf["viur.version"][1], 0))
-
-
-getVersion.exposed = True
 
 
 def canAccess(*args, **kwargs) -> bool:
@@ -149,21 +138,17 @@ def canAccess(*args, **kwargs) -> bool:
     return False
 
 
+@exposed
 def index(*args, **kwargs):
-    from viur.core.render import isAdminAvailable, isViAvailable
-    if not isViAvailable():
-        if isAdminAvailable():
-            # The Vi is not available, the admin however is, so redirect there
-            raise errors.Redirect("/admin")
+    if not conf["viur.instance.project_base_path"].joinpath("vi", "main.html").exists():
         raise errors.NotFound()
+
     if conf["viur.instance.is_dev_server"] or currentRequest.get().isSSLConnection:
         raise errors.Redirect("/vi/s/main.html")
+
     else:
         appVersion = currentRequest.get().request.host
         raise errors.Redirect("https://%s/vi/s/main.html" % appVersion)
-
-
-index.exposed = True
 
 
 @exposed
@@ -180,11 +165,9 @@ def get_settings():
 def _postProcessAppObj(obj):
     obj["skey"] = genSkey
     obj["timestamp"] = timestamp
-    obj["config"] = lambda *args, **kwargs: dumpConfig(conf["viur.mainApp"].vi)
-    obj["config"].exposed = True
+    obj["config"] = dumpConfig
     obj["settings"] = get_settings
-    obj["getStructure"] = lambda *args, **kwargs: getStructure(conf["viur.mainApp"].vi, *args, **kwargs)
-    obj["getStructure"].exposed = True
+    obj["getStructure"] = getStructure
     obj["canAccess"] = canAccess
     obj["setLanguage"] = setLanguage
     obj["getVersion"] = getVersion