+++ core/securityheaders.py
@@ -38,7 +38,7 @@
 
 
     ViUR also protects it's cookies by default (setting httponly, secure and samesite=lax). This can be changed by
-    setting the corresponding class-level variables on class:`GaeSession<viur.core.session.GaeSession>`.
+    setting the corresponding class-level variables on class:`Session<viur.core.session.Session>`.
 """
 
 from viur.core.config import conf
@@ -142,7 +142,7 @@ def extendCsp(additionalRules: dict = None, overrideRules: dict = None) -> None:
     assert additionalRules or overrideRules, "Either additionalRules or overrideRules must be given!"
     tmpDict = {}  # Copy the project-wide config in
     if conf["viur.security.contentSecurityPolicy"].get("enforce"):
-        tmpDict.update(conf["viur.security.contentSecurityPolicy"]["enforce"])
+        tmpDict.update({k: v[:] for k, v in conf["viur.security.contentSecurityPolicy"]["enforce"].items()})
     if overrideRules:  # Merge overrideRules
         for k, v in overrideRules.items():
             if v is None and k in tmpDict: