+++ core/skeleton.py
@@ -58,34 +58,39 @@ class MetaBaseSkel(type):
     __allowed_chars = string.ascii_letters + string.digits + "_"
 
     def __init__(cls, name, bases, dct):
-        boneMap = {}
-
-        def fill_bonemap_recursive(cls):
-            for base_cls in cls.__bases__:
-                if "__viurBaseSkeletonMarker__" in dir(base_cls):
-                    fill_bonemap_recursive(base_cls)
-
-            for key in cls.__dict__:
-                prop = getattr(cls, key)
-
-                if isinstance(prop, BaseBone):
-                    if not all([c in MetaBaseSkel.__allowed_chars for c in key]):
-                        raise AttributeError(f"Invalid bone name: {key!r} contains invalid characters")
-                    elif key in MetaBaseSkel.__reserved_keywords:
-                        raise AttributeError(f"Invalid bone name: {key!r} is reserved and cannot be used")
-
-                    boneMap[key] = prop
-
-                elif prop is None and key in boneMap:  # Allow removing a bone in a subclass by setting it to None
-                    del boneMap[key]
-
-        fill_bonemap_recursive(cls)
-        cls.__boneMap__ = boneMap
+        cls.__boneMap__ = MetaBaseSkel.generate_bonemap(cls)
 
         if not getSystemInitialized():
             MetaBaseSkel._allSkelClasses.add(cls)
 
         super(MetaBaseSkel, cls).__init__(name, bases, dct)
+
+    @staticmethod
+    def generate_bonemap(cls):
+        """
+        Recursively constructs a dict of bones from
+        """
+        map = {}
+
+        for base in cls.__bases__:
+            if "__viurBaseSkeletonMarker__" in dir(base):
+                map |= MetaBaseSkel.generate_bonemap(base)
+
+        for key in cls.__dict__:
+            prop = getattr(cls, key)
+
+            if isinstance(prop, BaseBone):
+                if not all([c in MetaBaseSkel.__allowed_chars for c in key]):
+                    raise AttributeError(f"Invalid bone name: {key!r} contains invalid characters")
+                elif key in MetaBaseSkel.__reserved_keywords:
+                    raise AttributeError(f"Invalid bone name: {key!r} is reserved and cannot be used")
+
+                map[key] = prop
+
+            elif prop is None and key in map:  # Allow removing a bone in a subclass by setting it to None
+                del map[key]
+
+        return map
 
 
 def skeletonByKind(kindName: str) -> Type[Skeleton]:
@@ -219,6 +224,9 @@ class SkeletonInstance:
     def __setattr__(self, key, value):
         if key in self.boneMap or isinstance(value, BaseBone):
             self.boneMap[key] = value
+        elif key == "renderPreparation":
+            super().__setattr__(key, value)
+            self.renderAccessedValues.clear()
         else:
             super().__setattr__(key, value)
 
@@ -405,7 +413,9 @@ class BaseSkeleton(object, metaclass=MetaBaseSkel):
 class MetaSkel(MetaBaseSkel):
     def __init__(cls, name, bases, dct):
         super(MetaSkel, cls).__init__(name, bases, dct)
-        relNewFileName = inspect.getfile(cls).replace(utils.projectBasePath, "").replace(utils.coreBasePath, "")
+        relNewFileName = inspect.getfile(cls) \
+            .replace(str(conf["viur.instance.project_base_path"]), "") \
+            .replace(str(conf["viur.instance.core_base_path"]), "")
 
         # Check if we have an abstract skeleton
         if cls.__name__.endswith("AbstractSkel"):
@@ -424,8 +434,9 @@ class MetaSkel(MetaBaseSkel):
                 cls.kindName = cls.__name__.lower()
         # Try to determine which skeleton definition takes precedence
         if cls.kindName and cls.kindName is not __undefindedC__ and cls.kindName in MetaBaseSkel._skelCache:
-            relOldFileName = inspect.getfile(MetaBaseSkel._skelCache[cls.kindName])\
-                .replace(utils.projectBasePath, "").replace(utils.coreBasePath,"")
+            relOldFileName = inspect.getfile(MetaBaseSkel._skelCache[cls.kindName]) \
+                .replace(str(conf["viur.instance.project_base_path"]), "") \
+                .replace(str(conf["viur.instance.core_base_path"]), "")
             idxOld = min(
                 [x for (x, y) in enumerate(conf["viur.skeleton.searchPath"]) if relOldFileName.startswith(y)] + [999])
             idxNew = min(